requirements.txt
boto3
Django>=3.2,<4.0
djangorestframework>=3.12.0,<4.0
psycopg2-binary>=2.8
python-magic
